requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.7.0
aiosignal==1.3.2
airportsdata==20241001
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
astor==0.8.1
async-timeout==5.0.1
attrs==24.3.0
autocommand==2.2.2
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
blake3==1.0.2
blis==1.1.0
blobfile==3.0.0
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cheroot==10.0.1
CherryPy==18.10.0
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
coloredlogs==15.0.1
colorful==0.6.0a1
compressed-tensors==0.9.1
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.0
cupy-cuda12x==13.3.0
cycler==0.12.1
cymem==2.0.10
datasets==3.2.0
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docx==0.2.4
einops==0.8.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
fastrlock==0.8.3
feedparser==6.0.11
filelock==3.16.1
flash-attn==2.7.3
flashinfer-python==0.2.0.post2
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
futures==3.0.5
genxword==2.2.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1rc0
google-api-python-client==2.162.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-genai==1.3.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
goslate==1.5.4
grpcio==1.70.0rc1
grpcio-status==1.70.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
interegular==0.3.3
isort==6.0.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
lightning-utilities==0.11.2
linkify-it-py==2.0.3
litellm==1.63.3
llvmlite==0.43.0
lm-format-enforcer==0.10.11
loralib==0.1.2
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
mistral_common==1.5.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
narwhals==1.25.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.61.0
opencensus==0.11.4
opencensus-context==0.2.dev0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
outlines==0.1.11
outlines_core==0.1.26
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
patsy==1.0.1
pdf2image==1.17.0
pdfminer.six==20240706
pdfminer3k==1.3.4
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
ply==3.11
portend==3.2.0
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycairo==1.27.0
pycountry==24.6.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
PyDictionary==2.0.1
Pygments==2.19.1
PyGObject==3.50.0
pyinflect==0.5.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.40.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
rsa==4.9
ruff==0.9.9
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.15.1
seaborn==0.13.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.6.0
setproctitle==1.3.3
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
starlette==0.41.3
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.9.0
tempora==5.8.0
tenacity==9.0.0
textual==1.0.0
thinc==8.3.3
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.14
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.4.0.post0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.49.0
transformers-stream-generator==0.0.5
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.1
vllm==0.7.3
wasabi==1.1.3
watchfiles==1.0.4
weasel==0.4.1
websockets==14.2
wrapt==1.17.2
xformers==0.0.28.post3
xgrammar==0.1.11
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
yq==3.4.3
zc.lockfile==3.0.post1
zipp==3.21.0
